requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
PyJWT==2.3.0
psycopg2-binary
SQLAlchemy==1.4.31
gunicorn
bcrypt
Flask-Limiter
flask-sqlalchemy
connexion
connexion[swagger-ui]
jsonpatch
Alembic
uuid
openalchemy